+++ crates/proof/challenge/Cargo.toml
@@ -68,11 +68,17 @@ url.workspace = true
 humantime.workspace = true
 axum = { workspace = true, features = ["http1", "tokio"] }
 
+# Test utilities (optional)
+alloy-rlp = { workspace = true, optional = true }
+alloy-trie = { workspace = true, optional = true }
+serde_json = { workspace = true, optional = true }
+async-trait = { workspace = true, optional = true }
+
 [dev-dependencies]
 rstest.workspace = true
-alloy-rlp.workspace = true
-alloy-trie.workspace = true
-serde_json.workspace = true
-async-trait.workspace = true
 reqwest = { workspace = true, features = ["json"] }
 tokio = { workspace = true, features = ["test-util"] }
+base-challenger = { path = ".", features = ["test-utils"] }
+
+[features]
+test-utils = [ "alloy-rlp", "alloy-trie", "async-trait", "serde_json" ]
