+++ crates/proof/challenge/tests/driver.rs
@@ -1,0 +1,499 @@
+//! Integration tests for the challenger [`Driver`] loop.
+
+use std::{
+    collections::HashMap,
+    sync::{Arc, Mutex, atomic::AtomicBool},
+    time::Duration,
+};
+
+use alloy_primitives::{Address, B256, Bytes};
+use base_challenger::{
+    ChallengeSubmitter, Driver, DriverConfig, GameScanner, OutputValidator, PendingProof,
+    ProofPhase, ScannerConfig,
+    test_utils::{
+        MockAggregateVerifier, MockDisputeGameFactory, MockGameState, MockL2Provider,
+        MockTxManager, MockZkProofProvider, addr, build_test_header_and_account, factory_game,
+        mock_state, receipt_with_status,
+    },
+};
+use base_enclave::output_root_v0;
+use base_proof_contracts::{AggregateVerifierClient, ContractError, GameAtIndex};
+use base_tx_manager::TxManagerError;
+use base_zk_client::ProofJobStatus;
+use tokio_util::sync::CancellationToken;
+
+/// Builds a test driver with the given mocks.
+fn test_driver(
+    factory: Arc<MockDisputeGameFactory>,
+    verifier: Arc<MockAggregateVerifier>,
+    l2_provider: Arc<MockL2Provider>,
+    zk_prover: Arc<MockZkProofProvider>,
+    tx_manager: MockTxManager,
+) -> Driver<MockL2Provider, MockZkProofProvider, MockTxManager> {
+    let scanner = GameScanner::new(
+        factory,
+        Arc::clone(&verifier) as Arc<dyn AggregateVerifierClient>,
+        ScannerConfig { lookback_games: 1000 },
+    );
+    let validator = OutputValidator::new(l2_provider);
+    let submitter = ChallengeSubmitter::new(tx_manager);
+
+    let config = DriverConfig {
+        poll_interval: Duration::from_millis(10),
+        cancel: CancellationToken::new(),
+        ready: Arc::new(AtomicBool::new(false)),
+    };
+
+    Driver::new(
+        config,
+        scanner,
+        validator,
+        zk_prover,
+        submitter,
+        verifier as Arc<dyn AggregateVerifierClient>,
+    )
+}
+
+fn default_zk_prover() -> Arc<MockZkProofProvider> {
+    Arc::new(MockZkProofProvider {
+        session_id: "test-session".to_string(),
+        proof_status: Mutex::new(ProofJobStatus::Created as i32),
+        receipt: Mutex::new(vec![]),
+    })
+}
+
+fn default_tx_manager() -> MockTxManager {
+    MockTxManager::new(Ok(receipt_with_status(true, B256::repeat_byte(0xAA))))
+}
+
+/// Builds the common L2, factory, and verifier mocks for an invalid-game
+/// scenario: starting=10, `l2_block=20`, interval=5, checkpoints at 15 and
+/// 20 with a correct root at 15 and a bogus root at 20 (invalid index 1).
+fn invalid_game_mocks()
+-> (Arc<MockL2Provider>, Arc<MockDisputeGameFactory>, Arc<MockAggregateVerifier>) {
+    let storage_hash = B256::repeat_byte(0xBB);
+    let (header_15, account_15) = build_test_header_and_account(15, storage_hash);
+    let root_15 = output_root_v0(&header_15, storage_hash);
+    let (header_20, account_20) = build_test_header_and_account(20, storage_hash);
+
+    let mut l2 = MockL2Provider::new();
+    l2.insert_block(15, header_15, account_15);
+    l2.insert_block(20, header_20, account_20);
+    let l2 = Arc::new(l2);
+
+    let factory = Arc::new(MockDisputeGameFactory { games: vec![factory_game(0, 1)] });
+    let mut verifier_games = HashMap::new();
+    verifier_games.insert(
+        addr(0),
+        MockGameState {
+            status: 0,
+            zk_prover: Address::ZERO,
+            tee_prover: Address::ZERO,
+            game_info: base_proof_contracts::GameInfo {
+                root_claim: B256::repeat_byte(0x01),
+                l2_block_number: 20,
+                parent_index: 0,
+            },
+            starting_block_number: 10,
+            intermediate_output_roots: vec![root_15, B256::repeat_byte(0xFF)],
+        },
+    );
+    let verifier = Arc::new(MockAggregateVerifier { games: verifier_games });
+
+    (l2, factory, verifier)
+}
+
+/// Builds a driver with a single pending `ReadyToSubmit` proof at `addr(0)`
+/// whose verifier reports the given `game_state`.
+fn driver_with_ready_proof(
+    game_state: MockGameState,
+) -> Driver<MockL2Provider, MockZkProofProvider, MockTxManager> {
+    let (l2, factory, _verifier) = invalid_game_mocks();
+    let verifier_games = HashMap::from([(addr(0), game_state)]);
+    let verifier = Arc::new(MockAggregateVerifier { games: verifier_games });
+    let mut driver = test_driver(factory, verifier, l2, default_zk_prover(), default_tx_manager());
+    driver.pending_proofs.insert(
+        addr(0),
+        PendingProof {
+            phase: ProofPhase::ReadyToSubmit {
+                proof_bytes: Bytes::from_static(&[0x01, 0xDE, 0xAD]),
+            },
+            invalid_index: 1,
+            expected_root: B256::repeat_byte(0xEE),
+        },
+    );
+    driver
+}
+
+#[tokio::test]
+async fn test_step_no_candidates() {
+    let factory = Arc::new(MockDisputeGameFactory { games: vec![] });
+    let verifier = Arc::new(MockAggregateVerifier { games: HashMap::new() });
+    let l2 = Arc::new(MockL2Provider::new());
+
+    let mut driver = test_driver(factory, verifier, l2, default_zk_prover(), default_tx_manager());
+
+    driver.step().await.unwrap();
+    // No error, no panic — empty scan is fine.
+}
+
+#[tokio::test]
+async fn test_step_valid_game_skipped() {
+    // Game with valid intermediate roots → no proof requested.
+    // We set up a game that will pass validation because intermediate_roots is empty
+    // and l2_block_number - starting_block_number < intermediate_block_interval
+    // so expected_count = 0 → trivially valid.
+    let factory = Arc::new(MockDisputeGameFactory { games: vec![factory_game(0, 1)] });
+    let mut verifier_games = HashMap::new();
+    verifier_games.insert(
+        addr(0),
+        MockGameState {
+            status: 0,
+            zk_prover: Address::ZERO,
+            tee_prover: Address::ZERO,
+            game_info: base_proof_contracts::GameInfo {
+                root_claim: B256::repeat_byte(0x01),
+                l2_block_number: 14,
+                parent_index: 0,
+            },
+            starting_block_number: 10,
+            intermediate_output_roots: vec![],
+        },
+    );
+    let verifier = Arc::new(MockAggregateVerifier { games: verifier_games });
+    let l2 = Arc::new(MockL2Provider::new());
+
+    // The ZK prover should NOT be called since the game is valid.
+    let zk = Arc::new(MockZkProofProvider {
+        session_id: "should-not-be-called".to_string(),
+        proof_status: Mutex::new(ProofJobStatus::Created as i32),
+        receipt: Mutex::new(vec![]),
+    });
+
+    let mut driver = test_driver(factory, verifier, l2, zk, default_tx_manager());
+
+    driver.step().await.unwrap();
+    // If the ZK prover were called, the test would still pass, but the game
+    // being valid means process_candidate returns early.
+}
+
+#[tokio::test]
+async fn test_step_validation_error_blocks_not_available() {
+    // Game with intermediate roots, but checkpoint blocks are unavailable.
+    // Validator returns BlockNotAvailable → process_candidate skips gracefully.
+    let factory = Arc::new(MockDisputeGameFactory { games: vec![factory_game(0, 1)] });
+    let mut verifier_games = HashMap::new();
+    verifier_games.insert(
+        addr(0),
+        MockGameState {
+            status: 0,
+            zk_prover: Address::ZERO,
+            tee_prover: Address::ZERO,
+            game_info: base_proof_contracts::GameInfo {
+                root_claim: B256::repeat_byte(0x01),
+                l2_block_number: 20,
+                parent_index: 0,
+            },
+            starting_block_number: 10,
+            intermediate_output_roots: vec![B256::repeat_byte(0xFF), B256::repeat_byte(0xEE)],
+        },
+    );
+    let verifier = Arc::new(MockAggregateVerifier { games: verifier_games });
+
+    // Checkpoint blocks are not available → validator returns BlockNotAvailable.
+    let mut l2 = MockL2Provider::new();
+    l2.error_blocks.push(15);
+    l2.error_blocks.push(20);
+    let l2 = Arc::new(l2);
+
+    let zk = Arc::new(MockZkProofProvider {
+        session_id: "test-session".to_string(),
+        proof_status: Mutex::new(ProofJobStatus::Created as i32),
+        receipt: Mutex::new(vec![]),
+    });
+
+    let mut driver = test_driver(factory, verifier, l2, zk, default_tx_manager());
+
+    // step succeeds — BlockNotAvailable causes process_candidate to skip
+    driver.step().await.unwrap();
+}
+
+#[tokio::test]
+async fn test_step_invalid_game_proof_succeeded() {
+    // Proof succeeds → nullification submitted.
+    let (l2, factory, verifier) = invalid_game_mocks();
+
+    let zk = Arc::new(MockZkProofProvider {
+        session_id: "proof-123".to_string(),
+        proof_status: Mutex::new(ProofJobStatus::Succeeded as i32),
+        receipt: Mutex::new(vec![0xDE, 0xAD]),
+    });
+
+    let tx_hash = B256::repeat_byte(0xCC);
+    let tx_manager = MockTxManager::new(Ok(receipt_with_status(true, tx_hash)));
+
+    let mut driver = test_driver(factory, verifier, l2, zk, tx_manager);
+
+    driver.step().await.unwrap();
+    // The tx_manager response was consumed → nullification was submitted.
+    // If it wasn't consumed, the next call would panic.
+}
+
+#[tokio::test]
+async fn test_step_invalid_game_proof_failed() {
+    // ZK prover returns Failed → no nullification submitted.
+    let (l2, factory, verifier) = invalid_game_mocks();
+
+    let zk = Arc::new(MockZkProofProvider {
+        session_id: "proof-fail".to_string(),
+        proof_status: Mutex::new(ProofJobStatus::Failed as i32),
+        receipt: Mutex::new(vec![]),
+    });
+
+    // tx_manager should NOT be called (proof failed → no submission)
+    let tx_manager = default_tx_manager();
+
+    let mut driver = test_driver(factory, verifier, l2, zk, tx_manager);
+
+    // step succeeds — proof failure is logged but not an error
+    driver.step().await.unwrap();
+}
+
+#[tokio::test]
+async fn test_step_validation_error_skipped() {
+    // Game where validator returns an error (e.g., BlockNotAvailable)
+    // → process_candidate logs and returns Ok.
+    let factory = Arc::new(MockDisputeGameFactory { games: vec![factory_game(0, 1)] });
+    let mut verifier_games = HashMap::new();
+    verifier_games.insert(
+        addr(0),
+        MockGameState {
+            status: 0,
+            zk_prover: Address::ZERO,
+            tee_prover: Address::ZERO,
+            game_info: base_proof_contracts::GameInfo {
+                root_claim: B256::repeat_byte(0x01),
+                l2_block_number: 20,
+                parent_index: 0,
+            },
+            starting_block_number: 10,
+            // 2 roots expected at interval=5, provide 2 so count matches
+            intermediate_output_roots: vec![B256::ZERO, B256::ZERO],
+        },
+    );
+    let verifier = Arc::new(MockAggregateVerifier { games: verifier_games });
+
+    // L2 provider has no blocks → validator returns BlockNotAvailable
+    let l2 = Arc::new(MockL2Provider::new());
+
+    let mut driver = test_driver(factory, verifier, l2, default_zk_prover(), default_tx_manager());
+
+    // step succeeds — validation error is skipped
+    driver.step().await.unwrap();
+}
+
+#[tokio::test]
+async fn test_step_scan_error_propagated() {
+    // Use ErrorOnIndexFactory where game_count succeeds but game_at_index
+    // errors. But scan itself catches per-game errors. To get scan to fail,
+    // we need game_count to fail, which requires a custom factory.
+
+    /// Factory that always fails on `game_count`.
+    #[derive(Debug)]
+    struct FailingFactory;
+
+    #[async_trait::async_trait]
+    impl base_proof_contracts::DisputeGameFactoryClient for FailingFactory {
+        async fn game_count(&self) -> Result<u64, ContractError> {
+            Err(ContractError::Validation("simulated game_count failure".into()))
+        }
+
+        async fn game_at_index(&self, _index: u64) -> Result<GameAtIndex, ContractError> {
+            unreachable!()
+        }
+
+        async fn init_bonds(
+            &self,
+            _game_type: u32,
+        ) -> Result<alloy_primitives::U256, ContractError> {
+            unreachable!()
+        }
+
+        async fn game_impls(&self, _game_type: u32) -> Result<Address, ContractError> {
+            unreachable!()
+        }
+    }
+
+    let factory = Arc::new(FailingFactory);
+    let verifier = Arc::new(MockAggregateVerifier { games: HashMap::new() });
+    let scanner = GameScanner::new(
+        factory,
+        Arc::clone(&verifier) as Arc<dyn AggregateVerifierClient>,
+        ScannerConfig { lookback_games: 1000 },
+    );
+
+    let l2 = Arc::new(MockL2Provider::new());
+    let validator = OutputValidator::new(l2);
+    let submitter = ChallengeSubmitter::new(default_tx_manager());
+
+    let config = DriverConfig {
+        poll_interval: Duration::from_millis(10),
+        cancel: CancellationToken::new(),
+        ready: Arc::new(AtomicBool::new(false)),
+    };
+
+    let mut driver = Driver::new(
+        config,
+        scanner,
+        validator,
+        default_zk_prover(),
+        submitter,
+        verifier as Arc<dyn AggregateVerifierClient>,
+    );
+
+    let result = driver.step().await;
+    assert!(result.is_err(), "scan error should propagate");
+}
+
+#[tokio::test]
+async fn test_step_pending_proof_skips_prove_block() {
+    // First step: proof initiated (status=Created, not ready).
+    // Second step: same game re-discovered → polls existing session,
+    // proof succeeds, nullification submitted.
+    let (l2, factory, verifier) = invalid_game_mocks();
+
+    let zk = Arc::new(MockZkProofProvider {
+        session_id: "pending-session".to_string(),
+        proof_status: Mutex::new(ProofJobStatus::Created as i32),
+        receipt: Mutex::new(vec![0xBE, 0xEF]),
+    });
+
+    let tx_hash = B256::repeat_byte(0xDD);
+    let tx_manager = MockTxManager::new(Ok(receipt_with_status(true, tx_hash)));
+
+    let mut driver = test_driver(factory, verifier, l2, Arc::clone(&zk), tx_manager);
+
+    // Step 1: proof is initiated but not ready (Created) → session stored.
+    driver.step().await.unwrap();
+    assert!(
+        driver.pending_proofs.contains_key(&addr(0)),
+        "session should be stored in pending_proofs"
+    );
+
+    // Simulate the proof completing before the next poll.
+    *zk.proof_status.lock().unwrap() = ProofJobStatus::Succeeded as i32;
+
+    // Step 2: same game re-discovered → polls existing session, proof succeeds,
+    // nullification submitted, session removed from pending_proofs.
+    driver.step().await.unwrap();
+    assert!(
+        !driver.pending_proofs.contains_key(&addr(0)),
+        "session should be removed after proof succeeded"
+    );
+}
+
+#[tokio::test]
+async fn test_step_nullification_failure_preserves_proof() {
+    // Proof succeeds on first step but nullification tx fails.
+    // The entry should stay in pending_proofs as ReadyToSubmit.
+    // On the next step the tx succeeds without re-proving.
+    let (l2, factory, verifier) = invalid_game_mocks();
+
+    let zk = Arc::new(MockZkProofProvider {
+        session_id: "proof-ok".to_string(),
+        proof_status: Mutex::new(ProofJobStatus::Succeeded as i32),
+        receipt: Mutex::new(vec![0xDE, 0xAD]),
+    });
+
+    // First tx call fails (NonceTooLow), second succeeds.
+    let tx_manager = MockTxManager::with_responses(vec![
+        Err(TxManagerError::NonceTooLow),
+        Ok(receipt_with_status(true, B256::repeat_byte(0xCC))),
+    ]);
+
+    let mut driver = test_driver(factory, verifier, l2, zk, tx_manager);
+
+    // Step 1: proof succeeds, but nullification tx fails.
+    // The error is swallowed by `step` (logged as a warning) because
+    // `poll_or_submit` returns Ok — only poll_pending_proofs logs it.
+    // But initiate_proof calls poll_or_submit directly, so its error
+    // propagates up through process_candidate → step logs it.
+    driver.step().await.unwrap();
+
+    // Entry must still be in pending_proofs as ReadyToSubmit.
+    let entry = driver.pending_proofs.get(&addr(0)).expect("proof should be preserved");
+    assert!(
+        matches!(entry.phase, ProofPhase::ReadyToSubmit { .. }),
+        "phase should be ReadyToSubmit after tx failure"
+    );
+
+    // Step 2: poll_pending_proofs re-submits, now the tx succeeds.
+    driver.step().await.unwrap();
+    assert!(
+        !driver.pending_proofs.contains_key(&addr(0)),
+        "entry should be removed after successful submission"
+    );
+}
+
+#[tokio::test]
+async fn test_poll_or_submit_drops_resolved_game() {
+    // Game has resolved (status=1 CHALLENGER_WINS) — driver should drop the
+    // pending proof without attempting submission.
+    let mut driver = driver_with_ready_proof(mock_state(1, Address::ZERO, 20));
+    driver.step().await.unwrap();
+    assert!(
+        !driver.pending_proofs.contains_key(&addr(0)),
+        "resolved game should be removed from pending_proofs"
+    );
+}
+
+#[tokio::test]
+async fn test_poll_or_submit_drops_already_challenged_game() {
+    // Game is still IN_PROGRESS but already challenged (zk_prover != ZERO)
+    // — driver should drop the pending proof.
+    let mut driver = driver_with_ready_proof(mock_state(0, Address::repeat_byte(0xCC), 20));
+    driver.step().await.unwrap();
+    assert!(
+        !driver.pending_proofs.contains_key(&addr(0)),
+        "already-challenged game should be removed from pending_proofs"
+    );
+}
+
+#[tokio::test]
+async fn test_run_cancellation() {
+    let factory = Arc::new(MockDisputeGameFactory { games: vec![] });
+    let verifier = Arc::new(MockAggregateVerifier { games: HashMap::new() });
+    let l2 = Arc::new(MockL2Provider::new());
+
+    let scanner = GameScanner::new(
+        Arc::clone(&factory) as Arc<dyn base_proof_contracts::DisputeGameFactoryClient>,
+        Arc::clone(&verifier) as Arc<dyn AggregateVerifierClient>,
+        ScannerConfig { lookback_games: 1000 },
+    );
+    let validator = OutputValidator::new(l2);
+    let submitter = ChallengeSubmitter::new(default_tx_manager());
+    let cancel = CancellationToken::new();
+
+    let config = DriverConfig {
+        poll_interval: Duration::from_secs(60), // long poll so it blocks
+        cancel: cancel.clone(),
+        ready: Arc::new(AtomicBool::new(false)),
+    };
+
+    let driver = Driver::new(
+        config,
+        scanner,
+        validator,
+        default_zk_prover(),
+        submitter,
+        verifier as Arc<dyn AggregateVerifierClient>,
+    );
+
+    // Cancel immediately
+    cancel.cancel();
+
+    // run() should return promptly
+    tokio::time::timeout(Duration::from_secs(2), driver.run())
+        .await
+        .expect("driver.run() should exit promptly after cancellation");
+}
